updated get checklist request to include limit
